requirements.txt
flask~=2.0.1
requests
Flask-SQLAlchemy
flask-login
Werkzeug~=2.0.1
SQLAlchemy~=1.4.23
boto3~=1.18.27
paho-mqtt~=1.5.1